Add some light correctness checks for mandatory fields and clean up

+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/messaging/AcceptedTypeEncoder.java
@@ -45,7 +45,7 @@ public final class AcceptedTypeEncoder extends AbstractDescribedListTypeEncoder<
     }
 
     @Override
-    public int getListEncoding(Accepted value) {
+    public byte getListEncoding(Accepted value) {
         return EncodingCodes.LIST0 & 0xff;
     }
