+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/messaging/DeleteOnNoLinksOrMessagesTypeEncoder.java
@@ -45,7 +45,7 @@ public final class DeleteOnNoLinksOrMessagesTypeEncoder extends AbstractDescribe
     }
 
     @Override
-    public int getListEncoding(DeleteOnNoLinksOrMessages value) {
+    public byte getListEncoding(DeleteOnNoLinksOrMessages value) {
         return EncodingCodes.LIST0 & 0xff;
     }
