+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/messaging/ModifiedTypeEncoder.java
@@ -62,7 +62,7 @@ public final class ModifiedTypeEncoder extends AbstractDescribedListTypeEncoder<
     }
 
     @Override
-    public int getListEncoding(Modified value) {
+    public byte getListEncoding(Modified value) {
         if (value.getMessageAnnotations() != null) {
             return EncodingCodes.LIST32;
         } else {
