+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/messaging/PropertiesTypeEncoder.java
@@ -104,7 +104,7 @@ public final class PropertiesTypeEncoder extends AbstractDescribedListTypeEncode
     }
 
     @Override
-    public int getListEncoding(Properties value) {
+    public byte getListEncoding(Properties value) {
         return EncodingCodes.LIST32;
     }
