+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/messaging/ReleasedTypeEncoder.java
@@ -45,7 +45,7 @@ public final class ReleasedTypeEncoder extends AbstractDescribedListTypeEncoder<
     }
 
     @Override
-    public int getListEncoding(Released value) {
+    public byte getListEncoding(Released value) {
         return EncodingCodes.LIST0;
     }
