+++ protonj2/src/main/java/org/apache/qpid/protonj2/codec/encoders/transport/ErrorConditionTypeEncoder.java
@@ -62,7 +62,7 @@ public final class ErrorConditionTypeEncoder extends AbstractDescribedListTypeEn
     }
 
     @Override
-    public int getListEncoding(ErrorCondition value) {
+    public byte getListEncoding(ErrorCondition value) {
         return EncodingCodes.LIST32;
     }
